benchmark: added Karmem benchmark
This change adds the Karmem benchmark. It have one trampoline, since Karmem
doesn't support some native features from CSharp/DotNet ( null and
Guid, to name a few). That also includes two types of benchmark, named
as "Reusable" and "Standard" (without suffix).

The "Standard" follows the same rules of any other benchmark. That will
allocate and destroy objects (or thrown then to GC) once done, without
any optimization.

The "Reusable" is reusing the allocated space. In real scenario, that happens when
have some resource pool, and objects are been reused instead of thrown to
garbage-collector. In the world of DotNet, that sounds like "ObjectPool". This
mode is how Karmem is supposed to use and was crafted to take some advantage of
such behavior, it also uses Span instead of Array, which is useful to share data
between memory pointers, instead of wired/network transfers.

Signed-off-by: Inkeliz <inkeliz@inkeliz.com>